feat(bet5): cut vector reads/writes over to pgvector + idempotent backfill
qdrant.ts now delegates to the pgvector store (proven at 1.000 ranking / identical filter parity),
so all ~dozen callers + the @dpf/db barrel keep their imports and transparently use Postgres. The
raw Qdrant REST client survives as qdrant-legacy.ts (rollback + backfill source, with a new
vector-inclusive scrollAllPoints). pgvector-backfill.ts copies each Qdrant collection into pgvector
once, idempotently (skips already-populated collections; preserves the numeric point ids). db
export-contract + translator tests green (11).

Process-Spine-Decision: BET-5 execution (plan Phase 1 cutover) — seam flip behind a proven parity gate; Qdrant retained one release for rollback.
BI: BI-A1E864A5
